Move markdown configuration to utils
